Update - UICollectionViewCompositionalLayout 적용

+++ SeSAC_Assignment_0131/SeSAC_Assignment_0131/Model/TVDetailsModel.swift
@@ -12,7 +12,7 @@ struct TVDetailsModel: Decodable {
     let name: String
     let numberOfEpisodes: Int
     let overview: String
-    let voteAverage: Float
+    let voteAverage: Double
     
     enum CodingKeys: String, CodingKey {
         case name, overview
